[Refactor] 파일 추가로 인한 디렉토리 위치 변경
- member.enums 패키지 추가
- exception.dto 패키지 추가

#63
