+++ src/main/java/com/appcenter/BJJ/domain/member/MemberService.java
@@ -2,6 +2,7 @@ package com.appcenter.BJJ.domain.member;
 
 import com.appcenter.BJJ.domain.member.domain.Member;
 import com.appcenter.BJJ.domain.member.dto.*;
+import com.appcenter.BJJ.domain.member.enums.MemberRole;
 import com.appcenter.BJJ.global.exception.CustomException;
 import com.appcenter.BJJ.global.exception.ErrorCode;
 import com.appcenter.BJJ.global.jwt.JwtProvider;
