+++ src/main/java/com/appcenter/BJJ/domain/member/enums/MemberRole.java
@@ -1,4 +1,4 @@
-package com.appcenter.BJJ.domain.member;
+package com.appcenter.BJJ.domain.member.enums;
 
 import lombok.Getter;
